Drawing with window offset
