+++ lua/bootstrap/options.lua
@@ -14,17 +14,13 @@ opt.ignorecase = true -- Ignore case
 opt.inccommand = "nosplit" -- preview incremental substitute
 opt.list = false -- Show some invisible characters (tabs...
 opt.mouse = "a" -- Enable mouse mode
-opt.number = false -- Print line number
 opt.pumblend = 10 -- Popup blend
 opt.pumheight = 10 -- Maximum number of entries in a popup
-opt.relativenumber = false -- Relative line numbers
-opt.scrolloff = 999 -- Lines of context
 opt.sessionoptions = { "buffers", "curdir", "tabpages", "winsize" }
 opt.shiftround = true -- Round indent
 opt.shiftwidth = 2 -- Size of an indent
 opt.shortmess:append({ W = true, I = true, c = true })
 opt.showmode = false -- Dont show mode since we have a statusline
-opt.sidescrolloff = 999 -- Columns of context
 opt.signcolumn = "yes" -- Always show the signcolumn, otherwise it would shift the text each time
 opt.smartcase = true -- Don't ignore case with capitals
 opt.smartindent = true -- Insert indents automatically
@@ -41,8 +37,6 @@ opt.wildmode = "longest:full,full" -- Command-line completion mode
 opt.winminwidth = 5 -- Minimum window width
 opt.wrap = false -- Disable line wrap
 opt.foldlevel = 99 -- Disable line wrap
-opt.cmdheight = 0 -- Disable command line at bottom
-opt.laststatus = 0
 
 if vim.fn.has("nvim-0.9.0") == 1 then
 	opt.splitkeep = "screen"
@@ -52,3 +46,11 @@ end
 -- Fix markdown indentation settings
 vim.g.markdown_recommended_style = 0
 
+opt.cmdheight = 0
+opt.laststatus = 0
+
+opt.number = true -- Print line number
+opt.relativenumber = false -- Relative line numbers
+
+opt.scrolloff = 0 -- Lines of context
+opt.sidescrolloff = 0 -- Columns of context
